pyproject.toml
[tool.poetry]
name = "my-python-scripts"
version = "0.1.0"
description = ""
authors = ["MarvinXu <272077995@qq.com>"]
readme = "README.md"
packages = [{include = "my_python_scripts"}]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^4.7.2"
requests = "^2.28.1"
loguru = "^0.6.0"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "my_python_scripts.__main__:run"
selenium = "my_python_scripts.selenium:run"
4k = "my_python_scripts.4ksj_selenium:run"

[[tool.poetry.source]]
name = "tsinghua"
default = true
url = "https://pypi.tuna.tsinghua.edu.cn/simple"